packages/constructor/pyproject.toml
[project]
name = "scp-constructor"
version = "0.1.0"
description = "Scan repositories for scp.yaml files and build architecture graphs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "neo4j>=5.0",
    "httpx>=0.25",
    "typer>=0.9",
    "rich>=13.0",
]

[project.scripts]
scp = "scp_constructor.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scp_constructor"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
